product price issue

+++ database/migrations/2026_03_12_103959_remove_price_from_products_table.php
@@ -9,11 +9,13 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+public function up(): void
 {
-    Schema::table('products', function (Blueprint $table) {
-        $table->dropColumn('price');
-    });
+    if (Schema::hasTable('products') && Schema::hasColumn('products', 'price')) {
+        Schema::table('products', function (Blueprint $table) {
+            $table->dropColumn('price');
+        });
+    }
 }
 
 public function down()
